# component/cron_task/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  httpserver.R:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    required:
    - code
    - msg
    type: object
info:
  contact:
    name: API Support
    url: https://github.com/hootrhino/rhilex
  description: rhilex Swagger API
  title: rhilex API
  version: "1.0"
paths:
...
  /crontask/results/page:
    get:
      consumes:
      - application/json
      parameters:
      - description: current
        in: query
        name: current
        type: string
      - description: size
        in: query
        name: size
        type: string
      - description: uuid
        in: query
        name: uuid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpserver.R'
      summary: 分页获取定时任务执行结果
      tags:
      - crontask
...
swagger: "2.0"